Code cleanup

+++ GolfGame1989/Environment.h
@@ -30,14 +30,14 @@ typedef struct tagBITMAPFILEHEADER {
 struct Environ
 {
     std::string                         name;
-    double                              airDensity;        // in kg/m^3
+    float                              airDensity;        // in kg/m^3
     std::string                         airDensityStr;
     DirectX::SimpleMath::Vector3        holePosition;
-    double                              gravity;           // in m/s^2
+    float                              gravity;           // in m/s^2
     std::string                         gravityStr;
-    double                              landingFrictionScale;
+    float                              landingFrictionScale;
     std::string                         landingFrictionScaleStr;
-    double                              landingHardnessScale;
+    float                              landingHardnessScale;
     std::string                         landingHardnessScaleStr;
     int                                 par;
     float                               scale;
@@ -82,25 +82,25 @@ class Environment
 public:
     Environment();
 
-    double GetAirDensity() const { return m_currentEnviron.airDensity; };
+    float GetAirDensity() const { return m_currentEnviron.airDensity; };
     std::string GetAirDensityString(const int aEnvironmentIndex) const { return m_environs[aEnvironmentIndex].airDensityStr; };
-    double GetAirDensity(const int aEnvironmentIndex) const { return m_environs[aEnvironmentIndex].airDensity; };
+    float GetAirDensity(const int aEnvironmentIndex) const { return m_environs[aEnvironmentIndex].airDensity; };
     DirectX::XMVECTORF32 GetEnvironColor() const { return m_currentEnviron.terrainColor; };
     std::string GetEnvironName(const int aEnvironmentIndex) const { return m_environs[aEnvironmentIndex].name; };
     std::vector<Fixture> GetFixtureBucket() { return m_fixtureBucket; };
 
-    double GetGravity() const { return m_currentEnviron.gravity; };
+    float GetGravity() const { return m_currentEnviron.gravity; };
     std::string GetGravityString(const int aEnvironmentIndex) const { return m_environs[aEnvironmentIndex].gravityStr; };
-    double GetGravity(const int aEnvironmentIndex) const { return m_environs[aEnvironmentIndex].gravity; };
+    float GetGravity(const int aEnvironmentIndex) const { return m_environs[aEnvironmentIndex].gravity; };
     DirectX::SimpleMath::Vector3 GetGroundHeight(DirectX::SimpleMath::Vector3 aPos);
     std::vector<DirectX::VertexPositionColor> GetFlagVertex() const { return m_flagVertex; };
     
     DirectX::SimpleMath::Vector3 GetHolePosition() const { return m_currentEnviron.holePosition; };
-    double GetHoleRadius() const { return m_holeRadius * m_currentEnviron.scale; };
+    float GetHoleRadius() const { return m_holeRadius * m_currentEnviron.scale; };
     std::vector<DirectX::VertexPositionColor> GetHoleVertex() const { return m_holeVertex; };
     std::vector<DirectX::VertexPositionNormalColor> GetHoleVertexTest1() const { return m_holeVertexTest1; };
-    double GetLandingHeight() const { return m_landingHeight; };
-    double GetLauchHeight() const { return m_launchHeight; };
+    float GetLandingHeight() const { return m_landingHeight; };
+    float GetLauchHeight() const { return m_launchHeight; };
     int GetNumerOfEnvirons() const { return m_environsAvailable; };
     int GetNumberOfEnvironSelectDisplayVariables() const { return m_environSelectDisplayDataPoints; };
     int GetPar() const { return m_currentEnviron.par; };
@@ -120,13 +120,13 @@ public:
     bool CheckTerrainTriangleHeight2(DirectX::XMFLOAT3& aPos, DirectX::XMFLOAT3 v0, DirectX::XMFLOAT3 v1, DirectX::XMFLOAT3 v2) const;
     bool CheckTerrainTriangleHeight3(DirectX::XMFLOAT3& aPos, DirectX::XMFLOAT3 v0, DirectX::XMFLOAT3 v1, DirectX::XMFLOAT3 v2) const;
     bool CheckTerrainTriangleHeight4(DirectX::XMFLOAT3& aPos, DirectX::XMFLOAT3 v0, DirectX::XMFLOAT3 v1, DirectX::XMFLOAT3 v2) const;
-    double GetWindDirection() const;
+    float GetWindDirection() const;
     DirectX::SimpleMath::Vector3 GetWindVector() const { return m_currentEnviron.wind; };
-    double GetWindX() const { return m_currentEnviron.wind.x; };
+    float GetWindX() const { return m_currentEnviron.wind.x; };
     std::string GetWindXString(const int aEnvironmentIndex) const { return m_environs[aEnvironmentIndex].windXStr; };
-    double GetWindY() const { return m_currentEnviron.wind.y; };
+    float GetWindY() const { return m_currentEnviron.wind.y; };
     std::string GetWindYString(const int aEnvironmentIndex) const { return m_environs[aEnvironmentIndex].windYStr; };
-    double GetWindZ() const { return m_currentEnviron.wind.z; };
+    float GetWindZ() const { return m_currentEnviron.wind.z; };
     std::string GetWindZString(const int aEnvironmentIndex) const { return m_environs[aEnvironmentIndex].windZStr; };
     bool InitializeTerrain();
     
@@ -146,8 +146,8 @@ private:
     void LoadFixtureBucket();
 
     bool LoadHeightMap();
-    void SetLandingHeight(double aLandingHeight);
-    void SetLauchHeight(double aLaunchHeight);
+    void SetLandingHeight(float aLandingHeight);
+    void SetLauchHeight(float aLaunchHeight);
     void ScaleTerrain();
 
     Environ                             m_currentEnviron;
@@ -162,12 +162,12 @@ private:
     std::vector<DirectX::VertexPositionColor> m_holeVertex;
     std::vector<DirectX::VertexPositionNormalColor> m_holeVertexTest1;
     const int                           m_holeResolution = 30;          // number of vertices used to draw hole circle
-    //const double                        m_holeRadius = 0.10795;       // Radius of the hole, future updates could include addition of "big cup" or "tiny cup" hole sizes
-    const double                        m_holeRadius = .50795;
+    //const float                        m_holeRadius = 0.10795;       // Radius of the hole, future updates could include addition of "big cup" or "tiny cup" hole sizes
+    const float                        m_holeRadius = .50795;
 
-    double                              m_landingHeight = 0.0;     // in meters
-    double                              m_launchHeight = 0.0;      // in meters
-    double                              m_testHeight = 0.0;        // in meters, for testing ground height functions
+    float                              m_landingHeight = 0.0;     // in meters
+    float                              m_launchHeight = 0.0;      // in meters
+    float                              m_testHeight = 0.0;        // in meters, for testing ground height functions
    
     std::vector<DirectX::VertexPositionNormal> m_heightMap;
     const float                         m_heightScale = 0.015;
@@ -178,18 +178,18 @@ private:
     std::vector<DirectX::VertexPositionNormal> m_terrainModel;
 
     // variables for bounce and roll functionality not yet implemented
-    double                              m_landingFriction;
-    double                              m_landingHardness;
-    double                              m_landingXslope;
-    double                              m_landingZslope;
+    float                              m_landingFriction;
+    float                              m_landingHardness;
+    float                              m_landingXslope;
+    float                              m_landingZslope;
 
     // min max consts
-    const double                        m_minAirDensity = 0.0;
-    const double                        m_maxAirDensity = 68.0; // just above the air density of Venus
-    const double                        m_minGravity = 0.1;
-    const double                        m_maxGravity = 28.0;    // approximate value for the mass of the sun
-    const double                        m_minMaxHeight = 450.0; // Launch & Landing min/max heights is just above the largest elevation change (>400 meters) of any real golf course which is the Extreme 19 in Limpopo Province South Africa
-    const double                        m_minMaxWind = 667.0;   // highest know wind speed on Neptune
+    const float                        m_minAirDensity = 0.0;
+    const float                        m_maxAirDensity = 68.0; // just above the air density of Venus
+    const float                        m_minGravity = 0.1;
+    const float                        m_maxGravity = 28.0;    // approximate value for the mass of the sun
+    const float                        m_minMaxHeight = 450.0; // Launch & Landing min/max heights is just above the largest elevation change (>400 meters) of any real golf course which is the Extreme 19 in Limpopo Province South Africa
+    const float                        m_minMaxWind = 667.0;   // highest know wind speed on Neptune
 
 
 };
